pdf_extractor_requirements.txt
pdfplumber>=0.10.3
pandas>=2.1.0
rich>=13.7.0
click>=8.1.7
openai>=1.12.0
python-dotenv>=1.0.0
tqdm>=4.66.1
loguru>=0.7.2 